+++ controllers/user.js
@@ -49,25 +49,33 @@ export const login = async (req, res, next) => {
 };
 
 export const updateUser = async (req, res, next) => {
-  try {
-    const user = await User.findByIdAndUpdate(
-      req.params.id,
-      { $set: req.body },
-      { new: true }
-    );
-    res.status(200).json(user);
-  } catch (err) {
-    next(err);
+  if (req.params.id === req.user.id) {
+    try {
+      const user = await User.findByIdAndUpdate(
+        req.params.id,
+        { $set: req.body },
+        { new: true }
+      );
+      res.status(200).json(user);
+    } catch (err) {
+      next(err);
+    }
+  } else {
+    res.status(400).json("You cannot update other users account");
   }
 };
 
 export const deleteUser = async (req, res, next) => {
-  try {
-    const user = await User.findByIdAndRemove(req.params.id);
-    res
-      .status(200)
-      .json(`user account ${user.firstName}, has been deleted succesfully`);
-  } catch (err) {
-    next(err);
+  if (req.params.id === req.user.id) {
+    try {
+      const user = await User.findByIdAndRemove(req.params.id);
+      res
+        .status(200)
+        .json(`user account ${user.firstName}, has been deleted succesfully`);
+    } catch (err) {
+      next(err);
+    }
+  } else {
+    res.status(400).json("You cannot delete other users account");
   }
 };
